// src/types/cache.enums.ts
export enum CacheKeys {
  PRODUCTS = 'products',
  PRODUCTS_LIST = 'products:list',
  PRODUCTS_DETAIL = 'products:detail',
  PRODUCTS_SEARCH = 'products:search',
  PRODUCTS_FILTER = 'products:filter',
  PRODUCTS_VARIATIONS = 'products:variations',
  BANNERS = 'banners',
  BANNERS_LIST = 'banners:list',
  CATEGORIES = 'categories',
  CATEGORIES_LIST = 'categories:list',
  CATEGORIES_DETAIL = 'categories:detail'
} 